Add DefaultConfiguration for builders, add egld transfer builders, fix after review, fix is_fungible().
