+++ erdpy_core/transaction_builders/__init__.py
@@ -2,12 +2,14 @@
 from erdpy_core.transaction_builders.base_builder import BaseBuilder
 from erdpy_core.transaction_builders.contract_builders import (
     ContractCallBuilder, ContractDeploymentBuilder, ContractUpgradeBuilder)
-from erdpy_core.transaction_builders.esdt_builders import (
-    ESDTIssueBuilder, ESDTNFTTransferBuilder, ESDTTransferBuilder,
+from erdpy_core.transaction_builders.esdt_builders import ESDTIssueBuilder
+from erdpy_core.transaction_builders.transfers_builders import (
+    EGLDTransferBuilder, ESDTNFTTransferBuilder, ESDTTransferBuilder,
     MultiESDTNFTTransferBuilder)
 
 __all__ = [
     "BaseBuilder",
     "ContractCallBuilder", "ContractDeploymentBuilder", "ContractUpgradeBuilder",
-    "ESDTIssueBuilder", "ESDTNFTTransferBuilder", "ESDTTransferBuilder", "MultiESDTNFTTransferBuilder"
+    "EGLDTransferBuilder", "ESDTNFTTransferBuilder", "ESDTTransferBuilder", "MultiESDTNFTTransferBuilder",
+    "ESDTIssueBuilder"
 ]
